controller - fight

+++ 2.2.1. spring_hibernate/src/main/java/hiber/entity/User.java
@@ -1,5 +1,6 @@
 package hiber.entity;
 
+
 import javax.persistence.*;
 
 @Entity
@@ -8,8 +9,8 @@ public class User {
 
    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
-   @Column (name = "id")
-   private Long id;
+   @Column(name = "id")
+   private int id;
 
    @Column(name = "name")
    private String name;
@@ -38,11 +39,11 @@ public class User {
       this.password = password;
    }
 
-   public Long getId() {
+   public int getId() {
       return id;
    }
 
-   public void setId(Long id) {
+   public void setId(int id) {
       this.id = id;
    }
